modulrize into src directory, move server code into seperate module
