+++ README.md
@@ -3,20 +3,23 @@
 voice clone via [eleven-labs](https://elevenlabs.io/) using existing youtube videos
 
 # Features
-- script `main.py` 
-- gunicorn server `gunicorn -w 1 -b 0.0.0.0:8889 -c gunicorn_config.py server:app --timeout 30`
+- script `src/main.py` 
+
+- gunicorn server `export PYTHONPATH=$PYTHONPATH:`pwd` && gunicorn -w 1 -b 0.0.0.0:8889 -c src/gunicorn_config.py src/server:app --timeout 30`
     - queue requests
     - handle request sequentially
 
 # Setup
-- Fill eleven labs API key in `audio_helper`
+
+- create `.env` template from `.example.env` with `cp .example.env .env`
+- Fill eleven labs API key in `.env`
 `You can view your xi-api-key using the 'Profile' tab on https://elevenlabs.io. Our API is experimental so all endpoints are subject to change.`
 - create a virtual env `python -m venv venv`
-- source env
+- source venv
 - install requirements `pip install -r requirements` tested using `python 3.10.10`
 
 ### Script
-- Usage `python main.py`
+- Usage `python src/main.py`
 - modify values in the `__name__ == "__main__"`
     - `video_urls` 
         - list of youtube videos to get audio from
@@ -26,7 +29,7 @@ voice clone via [eleven-labs](https://elevenlabs.io/) using existing youtube vid
         - name of file and folder clips will be saved to
 
 ### Server
-- Usage: `gunicorn -w 1 -b 0.0.0.0:8889 -c gunicorn_config.py server:app --timeout 30`
+- Usage: `export PYTHONPATH=$PYTHONPATH:`pwd` && gunicorn -w 1 -b 0.0.0.0:8889 -c src/gunicorn_config.py src/server:app --timeout 30`
 - Routes
     - `GET /q`
         - returns queue length
@@ -37,3 +40,4 @@ voice clone via [eleven-labs](https://elevenlabs.io/) using existing youtube vid
 ### TODOs
 [] Package structure
 [] Testing
+
