--- audio_helper.py
@@ -1,2 +1,0 @@
-from elevenlabs import set_api_key
-set_api_key("")
