+++ examples/2023-hormozi-example.json
@@ -1,0 +1,33 @@
+{
+    "name": "Alex hormozi (2023)",
+    "description": "Business personality youtuber, CEO Aquisition.com",
+    "youtube_links": [
+        "https://www.youtube.com/watch?v=60_7PU9JDIw", // I Convinced 10,000 People To Promote My Book For Me (For Free)
+        "https://www.youtube.com/watch?v=5cOwh-8scu8", // Make Money Online (Without Destroying Your Reputation)
+        "https://www.youtube.com/watch?v=7NMH1oAkgLY", // Giving Away Free Stuff Will Make You Rich
+        "https://www.youtube.com/watch?v=lJF__n_34ew", // My 12 Hour Work Day (in 15 Minutes)
+        "https://www.youtube.com/watch?v=Nh8Oc7ERdIU", // How to Get Ahead of 99% of People
+        "https://www.youtube.com/watch?v=zZyRg4Fzabk", // HUGE ANNOUNCEMENT
+        "https://www.youtube.com/watch?v=VPre_XMgKjs", // Why MrBeast Will be Worth $100 Billion
+        "https://www.youtube.com/watch?v=YFA8AS5Cu2w", // watch this if you're tired of being broke
+        "https://www.youtube.com/watch?v=1taVrxMFjaY&t=179s", // NO NEW FRIENDS (My "Extreme Views" on Friendship)
+        "https://www.youtube.com/watch?v=9gVdCR7W8o8", // The REAL Reason Your Business Isn't Growing
+        "https://www.youtube.com/watch?v=6ySRKgXBcO0&t=925s", // Day In The Life of Alex Hormozi
+        "https://www.youtube.com/watch?v=fxyhIXZ6Yog", // The Alex Hormozi Diet (REVEALED)
+        "https://www.youtube.com/watch?v=oTQPxPFROck&t=289s", // How to Build a LEGIT Online Course (2023)
+        "https://www.youtube.com/watch?v=KQuyQpFANpA", // I Solved Three $100,000,000 Business Problems
+        "https://www.youtube.com/watch?v=qLM5G7N3l3I", // 3 Ways to Do What You Love (and get wealthy too)
+        "https://www.youtube.com/watch?v=SYkwtqFoRcM", // How to go ALL IN on your side hustle..
+        "https://www.youtube.com/watch?v=zBZHWrvjD8Y", // How I Set Goals That ACTUALLY MAKE MONEY
+        "https://www.youtube.com/watch?v=LVM89ik-7Kw", // "I'm Broke, What Should I Sell?"
+        "https://www.youtube.com/watch?v=ueJg14gQLuc", // The SEASON OF NO (What it Takes to Win)
+        "https://www.youtube.com/watch?v=vZfatNSouDQ&t=179s", // Hardcore Business Lessons I Learned From A Dealer
+        "https://www.youtube.com/watch?v=zNJ5JzEJgyo", // Why You Shouldn't Copy Me
+        "https://www.youtube.com/watch?v=_PCCqqv2pig", // HOW TO GET WHAT YOU WANT (6 Proven Methods)
+        "https://www.youtube.com/watch?v=BHMeYaHEMpc", // How to Grow ANY Local Business (my framework)
+        "https://www.youtube.com/watch?v=ULGT0Qpglek&t=624s", // 17 Conversations That Made Me A Millionaire
+        "https://www.youtube.com/watch?v=-wnnwCqGeNc", // The #1 Reason Young People Stay Poor
+        "https://www.youtube.com/watch?v=0mqqbuM9sAk", // I Built 4 Businesses In A Row To Show It's Not Luck
+        "https://www.youtube.com/watch?v=ONV__y1z7MI", // How The World’s RICHEST Man Made His Money
+    ]
+}
