+++ examples/pre-2023-hormozi-example.json
@@ -1,0 +1,30 @@
+{
+    "name": "Alex hormozi (pre 2023)",
+    "description": "Business personality youtuber, CEO Aquisition.com",
+    "youtube_links": [
+        "https://www.youtube.com/watch?v=pWbSl7d0tEc", // https://www.youtube.com/watch?v=pWbSl7d0tEc
+        "https://www.youtube.com/watch?v=WRW7eY6K4Jo", // They lied to you: Why Punishment DOESN'T WORK..
+        "https://www.youtube.com/watch?v=FNJpJG-sSXM", // Biden's tax plan.....
+        "https://www.youtube.com/watch?v=-Pz0NpcILOY", // increased wealth EXPONENTIALLY
+        "https://www.youtube.com/watch?v=RBR4BwOVNvs", // solved everything
+        "https://www.youtube.com/watch?v=YaNX49ygr0I", // how to get what you want
+        "https://www.youtube.com/watch?v=Ec41cSp_tWc", // why direct response marketers get rich not wealth
+        "https://www.youtube.com/watch?v=kW2vDMAmlPI", // $100M CEO: How I know who to trust...
+        "https://www.youtube.com/watch?v=_gcqwupsza8", // I sold my company for $46,200,000 at age 32...7 LESSONS I learned..
+        "https://www.youtube.com/watch?v=o7R_K6LwKNk", // $100M CEO: "Why therapists failed me..." [language warning]
+        "https://www.youtube.com/watch?v=o7R_K6LwKNk", // $100M CEO: "Why therapists failed me..." [language warning]
+        "https://www.youtube.com/watch?v=F3NyhOtRpOE", // $100M CEO explains: How to CREATE a $100,000,000 product..[leaked training]
+        "https://www.youtube.com/watch?v=w5g0JiO7OdE", // I paid 0% tax on 56% of my income...here's how..
+        "https://www.youtube.com/watch?v=pWbSl7d0tEc", // I sold 8 businesses by age 32 [here’s how]
+        "https://www.youtube.com/watch?v=oi7bnS8uyJM", // 30 Minutes of sales training that will explode your business in 2022
+        "https://www.youtube.com/watch?v=3P1XjUvo1b4", // I cracked Starbucks RECURRING SALES model and built a FORTUNE...
+        "https://www.youtube.com/watch?v=_qspvJAq34M", // How I got 700 people to pay me $40,000 each...no bs..
+        "https://www.youtube.com/watch?v=kULFeI3LRYk", // I SPENT $100,000 on courses, so you don't have to...
+        "https://www.youtube.com/watch?v=C_SgvSvJZdk", // 8 Lessons Charlie Munger Taught Me To Build $112M Business
+        "https://www.youtube.com/watch?v=Q2VUuUGfpNY", // 6 Hacks To Make More Money [Increase LTV] Without More Customers
+        "https://www.youtube.com/watch?v=4LeHKDGmEIQ", // This is Why You're Not Happy - Eye Opening
+        "https://www.youtube.com/watch?v=Fy8XX8EuEnA", // I Made $31,714,980 in 2020 (Here's What I Learned)
+        "https://www.youtube.com/watch?v=cemduJKQl5w", // Recruiting and Hiring Tips - How to Hire The Best Employees
+        "https://www.youtube.com/watch?v=kULFeI3LRYk", // I SPENT $100,000 on courses, so you don't have to...
+    ]
+}
