+++ src/server/gunicorn_config.py
@@ -1,5 +1,5 @@
 def post_fork(server, worker):
-    from server import voice_generator, request_queue
+    from src.server.server import voice_generator, request_queue
     import threading
 
     threading.Thread(target=voice_generator, args=(request_queue,)).start()
